fix: restore missing constants.rs and fix all compilation errors
- Restored constants.rs from commit 3161095 with correct 1dcm³ voxel scaling
- Fixed CHUNK_SIZE = 50 (not 32) for 50×50×50 chunks
- Fixed block IDs to be u16 (not u32) throughout codebase
- Fixed MAX_BLOCK_DISTRIBUTIONS type mismatches (usize vs u32)
- Fixed shader include processing to actually process #include directives
- Fixed missing camera constants (DEFAULT_HEIGHT, CHUNK_SIZE_VOXELS)
- Updated all imports to use correct module paths
- Removed unnecessary _USIZE constant duplicates
- Fixed auto_wgsl macro array size parameters

The engine now compiles successfully. Root cause was missing constants.rs file
that was the single source of truth for all engine constants.

+++ src/gpu/soa/types.rs
@@ -6,7 +6,7 @@
 use encase::{ShaderType, ShaderSize, internal::WriteInto};
 use bytemuck::{Pod, Zeroable};
 use crate::gpu::types::terrain::{BlockDistribution, TerrainParams};
-use crate::gpu::constants::core::MAX_BLOCK_DISTRIBUTIONS;
+use crate::gpu::constants::MAX_BLOCK_DISTRIBUTIONS;
 use crate::gpu::automation::auto_wgsl::AutoWgsl;
 use crate::gpu::automation::auto_layout::AutoLayout;
 use std::marker::PhantomData;
@@ -270,7 +270,7 @@ mod tests {
         
         // Calculate expected size: count (4) + 5 arrays of MAX_BLOCK_DISTRIBUTIONS elements
         // Each array element is 4 bytes, so each array is MAX_BLOCK_DISTRIBUTIONS * 4
-        let expected_size = 4 + 5 * (MAX_BLOCK_DISTRIBUTIONS * 4);
+        let expected_size = 4 + 5 * (MAX_BLOCK_DISTRIBUTIONS as usize * 4);
         let expected_aligned = ((expected_size + 15) / 16) * 16; // Round up to 16-byte alignment
         
         println!("[BlockDistributionSOA] Expected size (unaligned): {} bytes", expected_size);
